Added addition $setting lookup to the getOption method to first look for a modUserSetting before looking for a modSystemSetting. This allows for a User Setting to then override a system setting.
Changed the naming convention of the cache_key to accomidate the prefix 'bigbrother-' for easy finding in cache folder along with the "bigbrother.account" value (user/system) then ending with an md5 hash of the request url as it was originally designed. This addition allows for each account to have its own cache and thus support mutiple accounts.

 Now by providing bigbrother.* at the UserSetting level one can support multiple Google Accounts (when multiple context are for multipel sites) or when Executive wants 90 days out while Manager get to see 30 days out.